requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
groq==0.9.0
ultralytics==8.2.0
torch==2.3.0
torchvision==0.18.0
prophet==1.1.5
ortools==9.10.4067
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
httpx==0.27.0
python-multipart==0.0.9
alembic==1.13.1
asyncpg==0.29.0
streamlit-extras==0.4.2
folium==0.16.0
streamlit-folium==0.20.0
